ready get and create

+++ src/main/java/co/istad/elearning/features/category/CategoryRepository.java
@@ -2,9 +2,13 @@ package co.istad.elearning.features.category;
 
 import co.istad.elearning.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category,Integer> {
+public interface CategoryRepository extends JpaRepository<Category,Long> {
 }
